+++ alg/holeelimination.cpp
@@ -166,7 +166,8 @@ int HoleEliminationParticle::adjFinishedNbrsDir() const {
 }
 
 
-HoleEliminationSystem::HoleEliminationSystem(uint numParticles, float holeProb) {
+HoleEliminationSystem::HoleEliminationSystem(uint numParticles,
+                                             double holeProb) {
   Q_ASSERT(numParticles > 0);
   Q_ASSERT(0 <= holeProb && holeProb <= 1);
 
@@ -200,7 +201,7 @@ HoleEliminationSystem::HoleEliminationSystem(uint numParticles, float holeProb) 
     occupied.insert(randomCandidate);
 
     // Add this candidate as a particle if not a hole.
-    if (randBool(1.0f - holeProb)) {
+    if (randBool(1.0 - holeProb)) {
       insert(new HoleEliminationParticle(randomCandidate, -1, randDir(), *this,
                                          HoleEliminationParticle::State::Inactive));
       ++numNonStaticParticles;
