+++ alg/infobjcoating.h
@@ -88,7 +88,7 @@ class InfObjCoatingSystem : public AmoebotSystem {
   // (#particles) and a hole probability. holeProb in [0,1] controls how "spread
   // out" the system is; closer to 0 is more compressed, closer to 1 is more
   // expanded.
-  InfObjCoatingSystem(uint numParticles = 100, float holeProb = 0.2);
+  InfObjCoatingSystem(uint numParticles = 100, double holeProb = 0.2);
 
   // Checks whether or not the system has completed infinite object coating (all
   // particles contracted and on the object.
