+++ alg/line.h
@@ -96,7 +96,7 @@ class LineSystem : public AmoebotSystem {
     // (#particles) and hole probability. holeProb in [0,1] controls how "spread
     // out" the system is; closer to 0 is more compressed, closer to 1 is more
     // expanded.
-    LineSystem(int numParticles = 200, float holeProb = 0.2);
+    LineSystem(int numParticles = 200, double holeProb = 0.2);
 
     // Checks whether or not the system's run of the line formation algorithm
     // has terminated (all particles in state Finish).
