+++ alg/ring.h
@@ -114,7 +114,7 @@ class RingSystem : public AmoebotSystem  {
   // (#particles) and hole probability. holeProb in [0,1] controls how "spread
   // out" the system is; closer to 0 is more compressed, closer to 1 is more
   // expanded.
-  RingSystem(int numParticles = 200, float holeProb = 0.2);
+  RingSystem(int numParticles = 200, double holeProb = 0.2);
 
   // Checks whether or not the system's run of the Ring formation algorithm
   // has terminated (all particles in state Finish).
